requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-debug-toolbar==4.1.0
django-filter==23.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
gamesdb==0.1.1
gevent==22.10.2
greenlet==2.0.2
idna==3.4
kombu==5.2.4
oauthlib==3.2.2
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
pypng==0.20220715.0
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
qrcode==7.4.2
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
zope.event==4.6
zope.interface==6.0
